BI-13524: Repackage tests.
