+++ src/test/java/uk/gov/companieshouse/alphabeticalcompanysearchconsumer/AlphabeticalCompanySearchConsumerApplicationTests.java
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ActiveProfiles;
+import uk.gov.companieshouse.alphabeticalcompanysearchconsumer.service.TestKafkaConfig;
 
 @SpringBootTest
 @Import(TestKafkaConfig.class)
